requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
fatsecret==0.2.3
Flask==1.1.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.0.0
pymongo==3.10.1
pytz==2019.3
rauth==0.7.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==0.16.1
WTForms==2.2.1
